requirements.txt
langchain>=0.2.0
langchain-groq>=0.1.0
pandas>=2.0.0
requests>=2.31.0
schedule>=1.2.0
python-dotenv>=1.0.0
